api/requirements.txt
fastapi
uvicorn
openai
pydantic
python-multipart
PyPDF2
numpy
python-dotenv